locks: bakery: invalidate stale lock data before releasing it
The PSCI framework uses bakery locks during CPU powerdown and powerup
sequences. During CPU power on, bakery_lock_get() is called with caches
disabled and bakery_lock_release() with caches enabled. During lock
acquisition, since caches are off, sometimes the invalidation at the end
of bakery_get_ticket() is not fully effective (presumably in the presence
of other masters concurrently attempting to read that line into the cache).
This can result in a stale copy of the bakery lock ticket in caches of
other cores.

This patch invalidates the stale copy of bakery ticket before the assert
check to avoid this scenario.

Signed-off-by: Krishna Sitaraman <ksitaraman@nvidia.com>
Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>